.proj changes, and fixing a shader bug
Changed the .proj and code in order not to copy the shaders to the .exe
directory, but use the shaders from their source directory so we don't
have to build if we change the shaders.

camera is slower

geometry shader was applying the default dumb answer to patches that
werent parallel to z=0, so i fixed that. Also realized it was parallel
to the wrong axis ( y) so i put it to be z as it should.